Move resize hook to its own file
